poison tiles added. proper lava tile behavior implemented. Force sliding works (airshot proof of concept), new font added, enemy deleted prop appears during battle sequence, explosion bug discovered and resolved

+++ BattleNetwork/bnCanodumb.cpp
@@ -123,5 +123,5 @@ const bool Canodumb::Hit(int _damage) {
 }
 
 const float Canodumb::GetHitHeight() const {
-  return hitHeight;
+  return (float)hitHeight;
 }
